backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
channels>=4.0
channels-redis>=4.2
daphne>=4.0
Pillow>=10.0
python-decouple>=3.8
